requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
groq
pydantic[email]
pydantic-settings
email-validator
aiofiles
